refactor: organize Dropdown folder
Move docs into /Documentation, unroll /Family
